# api-gateway/internal/service/handlers/metadata.go
package handler

import (
	"github.com/gin-gonic/gin"
)

// MetadataHandler handles video metadata operations
type MetadataHandler struct{}

// NewMetadataHandler creates a new metadata handler
func NewMetadataHandler() *MetadataHandler {
	return &MetadataHandler{}
}

// VideoResponse represents a video metadata response
type VideoResponse struct {
	ID          string `json:"id" example:"abc123"`
	Title       string `json:"title" example:"My Video Title"`
	Description string `json:"description" example:"Video description here"`
	UserID      string `json:"userId" example:"user123"`
	URL         string `json:"url" example:"https://example.com/videos/abc123"`
	Thumbnail   string `json:"thumbnail" example:"https://example.com/thumbnails/abc123.png"`
	Duration    int    `json:"duration" example:"180"`
	Views       int    `json:"views" example:"1000"`
	Status      string `json:"status" example:"published"`
	CreatedAt   string `json:"createdAt" example:"2023-01-01T12:00:00Z"`
	UpdatedAt   string `json:"updatedAt" example:"2023-01-01T12:00:00Z"`
}

// CreateVideoRequest represents a request to create a new video metadata
type CreateVideoRequest struct {
	Title       string `json:"title" example:"My Video Title" binding:"required"`
	Description string `json:"description" example:"Video description here"`
	UserID      string `json:"userId" example:"user123" binding:"required"`
	URL         string `json:"url" example:"https://example.com/videos/abc123"`
	Thumbnail   string `json:"thumbnail" example:"https://example.com/thumbnails/abc123.png"`
	Duration    int    `json:"duration" example:"180"`
}

// @Summary      Get all videos
// @Description  Get a list of all videos
// @Tags         metadata
// @Produce      json
// @Security     BearerAuth
// @Success      200  {array}   VideoResponse
// @Failure      401  {object}  map[string]interface{}  "Unauthorized"
// @Failure      500  {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/metadata/videos [get]
func (h *MetadataHandler) GetVideos() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the metadata-service
	return nil
}

// @Summary      Get video by ID
// @Description  Get a video's metadata by its ID
// @Tags         metadata
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Video ID"
// @Success      200  {object}  VideoResponse
// @Failure      401  {object}  map[string]interface{}  "Unauthorized"
// @Failure      404  {object}  map[string]interface{}  "Video not found"
// @Failure      500  {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/metadata/videos/{id} [get]
func (h *MetadataHandler) GetVideo() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the metadata-service
	return nil
}

// @Summary      Create video
// @Description  Create a new video metadata entry
// @Tags         metadata
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      CreateVideoRequest  true  "Video details"
// @Success      201      {object}  VideoResponse
// @Failure      400      {object}  map[string]interface{}  "Bad request"
// @Failure      401      {object}  map[string]interface{}  "Unauthorized"
// @Failure      500      {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/metadata/videos [post]
func (h *MetadataHandler) CreateVideo() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the metadata-service
	return nil
}

// @Summary      Update video
// @Description  Update a video's metadata
// @Tags         metadata
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        id       path      string              true  "Video ID"
// @Param        request  body      CreateVideoRequest  true  "Updated video details"
// @Success      200      {object}  VideoResponse
// @Failure      400      {object}  map[string]interface{}  "Bad request"
// @Failure      401      {object}  map[string]interface{}  "Unauthorized"
// @Failure      404      {object}  map[string]interface{}  "Video not found"
// @Failure      500      {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/metadata/videos/{id} [put]
func (h *MetadataHandler) UpdateVideo() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the metadata-service
	return nil
}

// @Summary      Delete video
// @Description  Delete a video's metadata
// @Tags         metadata
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Video ID"
// @Success      204  {object}  nil
// @Failure      401  {object}  map[string]interface{}  "Unauthorized"
// @Failure      404  {object}  map[string]interface{}  "Video not found"
// @Failure      500  {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/metadata/videos/{id} [delete]
func (h *MetadataHandler) DeleteVideo() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the metadata-service
	return nil
}