# api-gateway/internal/service/handlers/transcoder.go
package handler

import (
	"github.com/gin-gonic/gin"
)

// TranscoderHandler handles video transcoding operations
type TranscoderHandler struct{}

// NewTranscoderHandler creates a new transcoder handler
func NewTranscoderHandler() *TranscoderHandler {
	return &TranscoderHandler{}
}

// TranscodeJobRequest represents a request to create a new transcoding job
type TranscodeJobRequest struct {
	VideoID  string   `json:"videoId" example:"abc123" binding:"required"`
	Formats  []string `json:"formats" example:"480p,720p,1080p" binding:"required"`
	Priority int      `json:"priority" example:"1"`
}

// TranscodeJobResponse represents a transcoding job response
type TranscodeJobResponse struct {
	JobID     string   `json:"jobId" example:"job123"`
	VideoID   string   `json:"videoId" example:"abc123"`
	Status    string   `json:"status" example:"processing"`
	Progress  int      `json:"progress" example:"45"`
	Formats   []string `json:"formats" example:"480p,720p,1080p"`
	CreatedAt string   `json:"createdAt" example:"2023-01-01T12:00:00Z"`
	UpdatedAt string   `json:"updatedAt" example:"2023-01-01T12:05:00Z"`
}

// @Summary      Create a new transcoding job
// @Description  Submit a new video for transcoding
// @Tags         transcoder
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      TranscodeJobRequest   true  "Transcoding job details"
// @Success      201      {object}  TranscodeJobResponse  "Job created successfully"
// @Failure      400      {object}  map[string]interface{}  "Bad request"
// @Failure      401      {object}  map[string]interface{}  "Unauthorized"
// @Failure      500      {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/transcoder/jobs [post]
func (h *TranscoderHandler) CreateJob() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the transcoder-service
	return nil
}

// @Summary      Get job status
// @Description  Get the status of a transcoding job
// @Tags         transcoder
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Job ID"
// @Success      200  {object}  TranscodeJobResponse  "Job details"
// @Failure      401  {object}  map[string]interface{}  "Unauthorized"
// @Failure      404  {object}  map[string]interface{}  "Job not found"
// @Failure      500  {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/transcoder/jobs/{id} [get]
func (h *TranscoderHandler) GetJob() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the transcoder-service
	return nil
}