# api-gateway/internal/service/handlers/upload.go
package handler

import (
	"github.com/gin-gonic/gin"
)

// UploadHandler handles video upload operations
type UploadHandler struct{}

// NewUploadHandler creates a new upload handler
func NewUploadHandler() *UploadHandler {
	return &UploadHandler{}
}

// UploadVideoResponse represents an upload response
type UploadVideoResponse struct {
	ID       string `json:"id" example:"abc123"`
	Filename string `json:"filename" example:"my-video.mp4"`
	Size     int64  `json:"size" example:"1048576"`
	Status   string `json:"status" example:"processing"`
	URL      string `json:"url" example:"https://example.com/videos/abc123/my-video.mp4"`
}

// @Summary      Upload a video
// @Description  Upload a new video file
// @Tags         upload
// @Accept       multipart/form-data
// @Produce      json
// @Security     BearerAuth
// @Param        title formData string true "Video title"
// @Param        video formData file   true "Video file to upload"
// @Success      200   {object}  UploadVideoResponse
// @Failure      400   {object}  map[string]interface{}  "Bad request"
// @Failure      401   {object}  map[string]interface{}  "Unauthorized"
// @Failure      500   {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/upload/videos [post]
func (h *UploadHandler) UploadVideo() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the upload-service
	return nil
}

// @Summary      Process an uploaded video
// @Description  Start processing a previously uploaded video
// @Tags         upload
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      map[string]interface{}  true  "Process request"  schemas.ProcessVideoRequest
// @Success      200      {object}  map[string]interface{}  "Processing started"
// @Failure      400      {object}  map[string]interface{}  "Bad request"
// @Failure      401      {object}  map[string]interface{}  "Unauthorized"
// @Failure      404      {object}  map[string]interface{}  "Video not found"
// @Failure      500      {object}  map[string]interface{}  "Internal server error"
// @Router       /api/v1/upload/videos/process [post]
func (h *UploadHandler) ProcessVideo() gin.HandlerFunc {
	// This is just a placeholder for documentation
	// The actual implementation is in the upload-service
	return nil
}